NLogViewerTarget - Minor refactoring to remove use of internal properties
