+++ tests/NLog.UnitTests/RegressionTests.cs
@@ -41,25 +41,6 @@ namespace NLog.UnitTests
     public class RegressionTests : NLogTestBase
     {
         [Fact]
-        public void Bug3990StackOverflowWhenUsingNLogViewerTarget()
-        {
-            // this would fail because of stack overflow in the 
-            // constructor of NLogViewerTarget
-            var config = XmlLoggingConfiguration.CreateFromXmlString(@"
-<nlog>
-  <targets>
-    <target name='viewer' type='NLogViewer' address='udp://127.0.0.1:9999' />
-  </targets>
-  <rules>
-    <logger name='*' minlevel='Debug' writeTo='viewer' />
-  </rules>
-</nlog>");
-
-            var target = config.LoggingRules[0].Targets[0] as NLogViewerTarget;
-            Assert.NotNull(target);
-        }
-
-        [Fact]
         public void Bug4655UnableToReconfigureExistingLoggers()
         {
             var debugTarget1 = new DebugTarget();
